code restructur package
Aniket this is what we should follow I will  do more changes for now you can use the same struture
