--- PickShip-1/src/main/java/com/PickShip/CreateAccount/CreateAccount.java
@@ -1,34 +1,0 @@
-package com.PickShip.CreateAccount;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class CreateAccount {
-	@Id
-	@GeneratedValue(
-		    strategy= GenerationType.AUTO,
-		    generator="native"
-		)
-	Long customerId;
-	Long phoneNumber;
-	public Long getCustomerId() {
-		return customerId;
-	}
-	public void setCustomerId(Long customerId) {
-		this.customerId = customerId;
-	}
-	public Long getPhoneNumber() {
-		return phoneNumber;
-	}
-	public void setPhoneNumber(Long phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-	
-	@Override
-	public String toString() {
-		return "CreateAccount [customerId=" + customerId + ", phoneNumber=" + phoneNumber + "]";
-	}
-	
-}
