--- PickShip-1/src/main/java/com/PickShip/CreateAccountRepo/CreateAccountRepo.java
@@ -1,10 +1,0 @@
-package com.PickShip.CreateAccountRepo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.PickShip.CreateAccount.CreateAccount;
-
-public interface CreateAccountRepo extends JpaRepository<CreateAccount, Integer> {
-	
-	CreateAccount findByPhoneNumber(Long number);
-}
