--- PickShip-1/src/main/java/com/PickShip/PickShip1Application.java
@@ -1,13 +1,0 @@
-package com.PickShip;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PickShip1Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PickShip1Application.class, args);
-	}
-
-}
