+++ PickShip-1/src/main/java/com/pickship/PickShip1Application.java
@@ -1,0 +1,15 @@
+package com.pickship;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PickShip1Application
+{
+
+    public static void main(String[] args)
+    {
+	SpringApplication.run(PickShip1Application.class, args);
+    }
+
+}
