+++ PickShip-1/src/main/java/com/pickship/repo/CreateAccountRepo.java
@@ -1,0 +1,10 @@
+package com.pickship.repo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.pickship.entity.CreateAccount;
+
+public interface CreateAccountRepo extends JpaRepository<CreateAccount, Integer> {
+	
+	CreateAccount findByPhoneNumber(Long number);
+}
