+++ PickShip-1/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url = jdbc:mysql://localhost:3306/PickShip DB
+spring.datasource.url = jdbc:mysql://localhost:3306/pickship
 spring.datasource.username = root
 spring.datasource.password = root
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
